backend/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
factory_boy==3.3.1
Faker==33.1.0
inflection==0.5.1
pillow==11.0.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
pytz==2022.6
six==1.17.0
sqlparse==0.4.3
typing_extensions==4.12.2
tzdata==2022.6
